LayoutTests/fast/mediastream/microphone-change-while-muted-expected.txt

CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure


PASS Detection of missing capturing device triggers capture to fail even if capture was muted
